Introduce the event timeline
The event timeline is a representation of atomic events that occur when
playing a track. For now these include individual click sounds, and the
song's start point.

Using this, an event loop can be used similar to cwilso's metronome
which polls and schedules upcoming events without preempting the
entire timeline.

This has multiple benefits:
- Can modify individual events (e.g. change the frequency of certain
  clicks within a click track)
- Maximizes benefit of an event loop by having atomic events rather than
  batches
- Seeking to a particular point in the timeline is easily achievable