Introduce Player component to handle seeking and controls
The player component is responsible for scheduling events.
I may move away from having the song start as part of the event timeline
as it is handled significantly differently than click events.
The main difference is that seeking to middle of the track still
requires consumption of the song start event while past clicks can be
ignored. The only information needed is the song start time anyway.

I would also like to move away from the setTimeout event loop being on
the main UI thread for performance reasons, but it works for now.