Add Timeline component for a visual representation of events
The idea is for the events to dynamically move their location on screen
when modified, and display useful information such as duration of a
click track to see exactly where it will end.

Long term goal is to have this timeline be an editor as well where
individual click events can be modified (such as the sound played on the
click).