Naive implementation of preview mode
Preview mode allows the user to loop a section of the song automatically. This
commit implements a working version of this idea using setTimeout to reset
progress once the end time has been reached.

Ideally, this will be scheduled on another thread (Web Worker or WebAudio
thread) to unclog the main thread and remain responsive at all times.