Finally, scopes
I struggled big time with connecting to Spotify with the web playback
SDK after implementing the PKCE flow. I got everything to the point
where a valid access token was retrieved, but then a frustrating DRM
error messed with the actual song playback.

It turns out the default unscoped token is not enough (unsurprisingly).
This is probably something I could've deduced but I'm running on fumes.

Add in the 'streaming' scope along with a couple of others.