Use an environment variable for redirect URI
Previously, the redirect URI was defined in a src/ json file which, when
built, was inlined into a minified bundle js file. This makes it
frustrating to change when deploying in different environments.

The Vue CLI service lets me define environment variables that can be
replaced at build time to reflect an accurate redirect. This requires me
to downgrade the no-undef linting rule to a warning so the process.env.*
usages don't error out the build step.