Implement simple scene saving and loading
Name pending, but the ability to save scenes (a.k.a the song with all
click tracks) is an important feature, especially when the scenes get
more and more complicated.

The loading flow requires a new authorization scope to change the
currently playing song. The behavior is a bit frustrating right now
where the song plays for a fraction of a second and then I manually
pause since there is no change-but-don't-play Spotify API call.
Preferably, I'll mute playback during the switch.

Add in basic functionality that saves and loads from localStorage to
persist between sessions.