Begin migration to Vuex for state management
This is already a nice improvement, even with just the starting point of
storing the tokens and API. The devtools timeline seems like it'll be
huge to see the ordering in which state changes were effected.

The single central store also makes it convenient to see the entire
state without traversing down the component tree, and if I end up using
Vuex modules, the store will be nicely grouped when observing it.