Extract SceneSaver component
This component is responsible for saving scenes in state as well as
persisting and loading from the browser's localStorage across sessions.

localStorage interaction currently only happens upon logging in to the
website (loading) and closing the browser tab, triggering the
`beforeunload` event (persisting). If the persistence before unload is
unstable and inconsistent, I may move over to persistence on every scene
save to guarantee an up-to-date localStorage representation.