Begin use of requestAnimationFrame for scheduling
Request animation frame can be expected to give us < 16ms (>= 60FPS) of
latency for scheduling events. It definitely makes sense to use this
over setTimeout since there is a stronger guarantee that
requestAnimationFrame's timestamp is precise to the millisecond and it
won't be queued behind other messages on the message queue.

This change introduces the requestAnimationFrame loop for setting up the
Spotify song (no preview mode).