Add preview duration rule to prevent undefined behavior
When the preview duration is sufficiently small, the behavior can become
unpredictable as there is extremely frequent restarting of both the
event loop and the Spotify playback.

Previewing anything under 3 seconds is not expected to be desired by a
user, so we can explicitly disallow it in the v-text-field rules.