Fix seeking in preview mode
Preview mode seeking regressed with the change to one-way binding of the
seek slider. There were dangling references to the no-longer-existent
`sliderValue` state. Address this by using `seekOffsetSeconds` to
reset the timeline in each preview loop iteration.

Pausing and continuing in preview mode is currently undesirable: the seek offset is
updated to the pause point, so the loop section is not preserved
(shifted to the right). Ideally, the user can resume from the middle of
a loop and it will reset back to the originally set seek time. This will
involve some non-trivial logic to setTimeout the preview loop on a
shorter delay for the first iteration.