Simplify ClickTrack layout: append remove button to slider
Instead of treating the remove button as a separate v-col, add it to the
append slot of the start time slider to achieve the same effect.

I would like the remove to be a FAB at the top-right of each click track
but 'absolute top right' positions all of the click tracks' remove
buttons top-right of the card component encapsulating the
TimelineEditor.