Only update slider bound data on release
Previously, the song start and click start start sliders were two-way
bound to their corresponding reactive state. This means that during the
entire duration that the user holds onto the slider, the reactive state
was being updated and propagated to the listeners of these data.

Address this by only updating the reactive state when the user releases
the slider.