Nudge every click by 100ms to avoid crackling on starting click
Previously, a click track scheduled for t=0 would crackle harshly on the
first click sound. This was due to the small amount of time passed since
starting the event loop, and the fact that the click scheduling behavior
looks behind to scoop up any very recently missed clicks. In practice,
this issue can arise when seeking to a point where a click lines up, not
just the very beginning of the track.

To address the crackling, nudge all clicks by a fixed amount when
scheduling them in the audio context. I have chosen 100ms for now, which
should be more than enough. Also, the click scheduling look-behind is
now half of the nudge value. This way a click with a -50ms computed time
(relative to the current audio context time) would be scheduled for
50ms after nudging. Any click earlier than -50ms would be ignored as it
is outside of the look-behind window.

I have not nudged the playback of the Spotify song yet, as it seems
extraneous. BUT, if this becomes an issue with the purported start times
not aligning with what the user expects, then the song can be nudged in
the future.