Remove usages of v-slider persistent-hint attribute
Vuetify's slider performance takes a huge hit when using the
persistent-hint attribute. This was being used in the song and click
start time sliders to display the times.

Use label (like the seek bar) to communicate these rather than hints to
improve performance. It is a bit cramped but the UI can be improved in
the future.